+++ src/Symfony/Component/Form/Button.php
@@ -347,6 +347,15 @@ class Button implements \IteratorAggregate, FormInterface
     /**
      * Unsupported method.
      *
+     * @return null Always returns null
+     */
+    public function getTransformationFailure()
+    {
+    }
+
+    /**
+     * Unsupported method.
+     *
      * @throws BadMethodCallException
      */
     public function initialize()