+++ src/Symfony/Component/Form/Tests/Extension/DataCollector/FormDataExtractorTest.php
@@ -319,7 +319,7 @@ class FormDataExtractorTest extends \PHPUnit_Framework_TestCase
                 'norm' => "'Foobar'",
             ),
             'errors' => array(
-                array('message' => 'Invalid!', 'origin' => null, 'cause' => null),
+                array('message' => 'Invalid!', 'origin' => null, 'trace' => array()),
             ),
             'synchronized' => 'true',
         ), $this->dataExtractor->extractSubmittedData($form));
@@ -340,7 +340,7 @@ class FormDataExtractorTest extends \PHPUnit_Framework_TestCase
                 'norm' => "'Foobar'",
             ),
             'errors' => array(
-                array('message' => 'Invalid!', 'origin' => spl_object_hash($form), 'cause' => null),
+                array('message' => 'Invalid!', 'origin' => spl_object_hash($form), 'trace' => array()),
             ),
             'synchronized' => 'true',
         ), $this->dataExtractor->extractSubmittedData($form));
@@ -360,7 +360,12 @@ class FormDataExtractorTest extends \PHPUnit_Framework_TestCase
                 'norm' => "'Foobar'",
             ),
             'errors' => array(
-                array('message' => 'Invalid!', 'origin' => null, 'cause' => 'object(Exception)'),
+                array('message' => 'Invalid!', 'origin' => null, 'trace' => array(
+                    array(
+                        'class' => "'Exception'",
+                        'message' => "''",
+                    ),
+                )),
             ),
             'synchronized' => 'true',
         ), $this->dataExtractor->extractSubmittedData($form));